TRIB-87/88
Added changeSet to add Notifications table and notification_type table.
Added changeSet to add foriegn key linking the tables.
Added changeSet to add mock data for testing in both tables.
Verified that the tables,rows, and values were added to the Mysql database and that the backend was able to run without issues.